Azure/function-app/requirements.txt
azure-functions
Flask==1.1.1
azure-storage-blob==12.3.0
python-dateutil==2.8.1
azure-eventhub==5.0.0
azure-eventhub-checkpointstoreblob==1.0.0
azure-cosmosdb-table==1.0.6